assignment/assignment_for_RTK_using_backend_drf/tourmate/requirements - Copy.txt
Django==1.11.17
djangorestframework==3.8.2
greenlet==3.0.3
importlib-metadata==6.7.0
Pillow==9.5.0
pytz==2024.1
SQLAlchemy==2.0.31
typing-extensions==4.7.1
zipp==3.15.0
